I had trouble sleeping during the night and it is all because of the idea. I cannot reason out whether ensemble linking via reparameterization is a good idea. Well, if it fails I am going to look into randomized priors and such, but...
It is a pain. I've been thinking about this during the night, but Spiral is not the right language to do experimentation in. A language suited for that should have automatic type inference, great IDE support and fast compilation. Spiral is more of a language where you put everything together.

https://web.facebook.com/nipsfoundation/videos/2120856364798049/?_rdc=1&_rdr
Reproducible, Reusable, and Robust Reinforcement Learning

Let me watch this video for a bit and then I'll read the papers. Apart from Osband's papers, I'll look up citations for the ensembling paper that I read yesterday.

This is effort is going to be the last time. Bayesian inference has its role, but obviously I am not going to be running MCMC. If ensembling fails me then the quest will have failed. I absolutely need to start from some position of strength rather than the position of absolute weakness I am in now.

If I could have an RL algorithm that is not absolutely awful - one that is thorough in its exploration then I could figure something out as I go along. Otherwise, forget Spiral. I might as well start looking for a job.

I am really groggy at the moment, so let me watch the video.