"9:05am. Forget the video.
9:55am. I watched Spyce instead. Let me just start with the papers.

https://arxiv.org/abs/1809.04270
Rapid Training of Very Large Ensembles of Diverse Neural Networks

https://arxiv.org/abs/1807.01477
Diversity in Machine Learning

https://arxiv.org/abs/1804.00382
Attention-based Ensemble for Deep Metric Learning

Forget these for now. Let me go with randomized value functions paper once more from the top...actually, no, there were two papers that I did not get to yesterday. Let me start with the one by Osband and then I'll take a look at the GP one.

10:10am. http://mlg.eng.cam.ac.uk/yarin/blog_2248.html
Uncertainty in Deep Learning

I really should read this guy's thesis at some point. And point is close to here. Ultimately, I am not sure if there is a point in reading all this stuff now rather than diving into coding since I know what I need to do, but aiming for a broad coverage might be useful down the line.

That integrating over models is a good way to model uncertainty is not at all obvious. I mean, before I started studying this I hypothesized that there might be some kind of method to do it directly. I thought there might be a way to make dual numbers that represent uncertainty in the inputs and have the weights modify them. In the absence of that I suppose ensembles will do as well.

10:20am. https://arxiv.org/abs/1507.00300
Bootstrapped Thompson Sampling and Deep Exploration

https://projecteuclid.org/download/pdf_1/euclid.aos/1176345338
The Bayesian Bootstrap

10:25am. 4/5. Actually, there are some simulations in the appendix. Ok, this explains why they like the bootstrap over the ensemble.

And what the call smoothing is merely adding noise to the outputs. This quite succinctly provides an answer to whether should I add noise to the outputs. I very well should. This is a very strong point in favor of ensemble linking via reparameterization trick.

10:35am. Let me go for the GP paper next.

10:40am. Ah, crap I get it. I know what the problem is with the intrinsic motivation idea of mine. The issue with ensemble linking is that it will also model the data uncertainty. In essence that means that moves with high data uncertainty will motivate the agent. This would in essence be a gambling compulsion.

I think I'll just get rid of the intrinsic motivation part then. It might as well be that way, posterior sampling methods do not need intrinsic rewards anyway. That idea was something I had before I started studying this anyway.

11:10am. One more thing...

https://www.semanticscholar.org/paper/Risk-versus-Uncertainty-in-Deep-Learning-%3A-Bayes-%2C-Osband/dde4b95be20a160253a6cc9ecd75492a13d60c10
Risk versus Uncertainty in Deep Learning - Bayes, Bootstrap and the Dangers of Dropout

Let me plug this paper again. On page 4 there are some comparisons between bootstrapping and ensembling and objectively bootstrapping + smoothing gives the best uncertainty estimates. But now I understand that ensemble linking takes into account the data uncertainty, that actually casts shade on the bootstrap results. I think that the naive ensemble is the one which in fact measures the model uncertainty best on the datasets. On figure 3 for example, the mean is correct, the data noise is just not taken into account.

Why is it smoothing with variance 1/4? Ensemble linking solves that problem.

Just what the hell is bootstrapping measuring? It certainly should not be measuring data uncertainty.

Let me finish reading the GP paper and I will take a break.

11:35am. I skimmed the bootstrap paper as well. Let me stop here.

I guess in the end there is a decent bit of terror in programming I never accounted for. I never really thought about the other side of algorithmic reality. Because algorithms are truly unique that means that if I missed the right one, I will truly be lost.

If anchored ensemble linking via reparameterization trick sampling fails as a deep exploration method...and not just that, if it does not turn out to be the best of all the alternatives then I will truly be lost. I will be cast into the darkness once more. There are no substitutes for it. Not as far I can see. And the loss of opportunity will be something that will be hard to recover from.

Today I had the insight that I do not need intrinsic rewards, but that only served to chisel out what the method really should be.

I need to go through that paper that has been compared against in the achored ensemble paper. Then I'll take a look at Yarin's thesis.

I'll do that after the break."