"2:05pm. "For regression problems, NNs usually output a single value say µ(x) and the parameters are optimized to minimize the mean squared error (MSE) on the training set, given by ... However, the MSE does not capture predictive uncertainty. Following [47], we use a network that outputs two values in the final layer, corresponding to the predicted mean µ(x) and variance σ^2 (x) > 0. By treating the observed value as a sample from a (heteroscedastic) Gaussian distribution with the predicted mean and variance, we minimize the negative log-likelihood criterion."
https://ieeexplore.ieee.org/document/374138
Estimating the mean and variance of the target probability distribution

Here is the paper that is referred.

The pdf is behind a paywall. I'll track it down later. It seems in this paper they come close, but not quite exactly to my idea.

2:25pm. Forget this paper. The anchored ensemble paper is superior to the one in the paper I just read.

I am going back to the thesis by Yarin."