"6:30pm. 119/174. The stuff on Pilco is interesting. So they can do it using deep nets.
6:50pm. "In section 1.2 we discussed the different types of uncertainty encountered in Bayesian modelling: epistemic uncertainty which captures our ignorance about the models most suitable to explain our data, aleatoric uncertainty which captures noise inherent in the environment, and predictive uncertainty which conveys the model’s uncertainty in its output."

So ensemble linking via reparameterization sampling is in fact a way of estimating the predictive uncertainty.

7:15pm. https://arxiv.org/abs/1605.06465
Swapout - Learning an ensemble of deep architectures

Found this as one of the references. Let me take a peek and I will call it a day. I am finally done with the thesis. A lot of it was on dropout, but I don't like dropout in particular. I am more convinced by Osband that is estimating the wrong kind of uncertainty.

7:25pm. Gimmick paper. I am done for today.

I am starting to get a bit hopeful about all of this. The reason why Osband might not have had much luck with ensembles is because Q learning does not propagate uncertainty backwards through the cost function. It is just not possible for it to unless he specifically tried sampling the values and then taking the max, but the chance he did something like that is zero. It would also be necessary to weight it by the policy probability because of the whole delusion thing introduced otherwise. Off policy Q learning is fundamentally broken in its current form.

The reason bootstrap works better is because it introduces extra model uncertainty indirectly. The non-bootstrapped lower layers learn as well and I seem to recall that trained net when retrained on a fresh random layer in the head learns much more quickly.

This idea that I have will have will in fact work as the thorough solution to deep exploration leaving the issue generalization to be tackled next.

The order I came to this idea is weird. Usually one would think of local contraction to the mean using the square error rather than the more complicated sampling scheme, but all is well that ends well.

The thing I had in mind a week ago had various issues - lack of symmetry and composability with various possible architectures, but what I have in mind now is extremely broad. It completely generalizes intrinsic motivation through noise propagation and uncertainty estimation through ensemble linking. It has an answer to a wide variety of possible objections and has a deep elegance to it that appeals to me as a programmer.

It is good that for the past few days I've been essentially wasting time studying. It really allows me to refine my mental model while accumulating knowledge on the side. I am definitely following a good policy.

7:40pm. Tomorrow I will dive again into Osband's papers and after that rest a bit while I compose the monthly update and the yearly review. Since every day I've had a new insight with regard to my plan, it would be good to let it all settle down before I proceed to carve it into code.

It seems every day for the past 3.5 years had been full of surprises. I did not think I would be coming back into deep RL so quickly after throwing in the towel. Osband's work and the anchored ensemble papers really are deeply inspirational to me. I will be able to do this thanks to them.

I did a good think by delaying my entry into deep RL through my work on Spiral.

If I can move from having a completely broken algorithm that cannot do exploration properly to one that can do decently given its capacities, I will reach the entry point. It will give me the capability to wait for the opportunity that will come after this one and the power to reach it.

After so many failed experiments in the past months I really feel the reaper's scythe closing in on my neck. Maybe I should be grateful to the reaper for without its shadow looming over me, I could never put in as much effort as I did.

Really, anyone could come up with the ensemble linking idea with the knowledge that I have, but as little as a week ago, I had no idea that priors are those that determine what a function predicts on out of sample data. This is something that should have been obvious to me, but I simply never thought about it. Now that piece of knowledge is incredibly important to me."