"9am. I am up. I had trouble sleeping again, but this time is because I've had some powerful insights during the night. Rather than sampling, I realized that for propagating uncertainty through the value function I'd be better off integrating. There should be a squared error for univariate Gaussians.
9:20am. Let me chill a little. I am so groggy that I have the focus of a moth right now.

9:40am. https://en.wikipedia.org/wiki/Sum_of_normally_distributed_random_variables

So the variances get summed in addition. Interesting. What about multiplication?

https://en.wikipedia.org/wiki/Propagation_of_uncertainty

9:50am. The example formulas on this wiki page are immensely helpful. They are everything I need to implement the ideas I have in mind.

10:20am. I am thinking about correlations when doing uncertainty propagation.

```
bck = inl x out -> half / out
```

When `out` is zero here let me just set the gradient to zero.

```
    inl sqrt =
        unary {
            fwd = sqrt
            bck = inl x out -> if out = zero then zero else half / out
            }
```

Let me do it like this. I want to be able to propagate uncertainties for covariances even when the variable might have zero variance. To calculate the covariance I will have to take the square root of the variance. This is right in the limit as out approaches zero anyway.

Though assuming that variances that come from the model are correlated is a design choice. I actually do not know if they really are. This is something I would not have to worry about when sampling.

But I'll just go with the assumptions. I am really worried about the variance should I propagate uncertainties through sampling.

10:30am. Come to think of it, it is really a miracle that none of the RL algorithms take advantage of uncertainty propagation. That is near 2019 in a nutshell for you.

10:35am. The plan really is starting to come together. All of this has enough explosive power to open a rift to the next stage now.

10:40am. http://web.mit.edu/fluids-modules/www/exper_techniques/2.Propagation_of_Uncertaint.pdf
Propagation of Uncertainty through Mathematical Operations

Maybe I should study some of the literature on this. I am a bit confused why all the wavy equal signs."