"10:15am. Ah, I am just wasting time. I am not even reading manga. Let me finally start with the papers. Let me finally finish the randomized value function paper.
10:50am. 20/55. This stuff about noise in the algorithm having to be consistent is confusing me. I get the feeling that I am missing just one tiny bit. I get the sense that I understand noise propagation, but...

"It is worth commenting on a couple subtle properties of learn grlsvi. First, given θh and H, θh+1 is sampled from a Gaussian distribution. However, given the inputs to gaussian rlsvi, the output ˜θH is not Gaussian. This is because θh+1 depends nonlinearly on θh due to the maximization over actions in Line 9 of the algorithm. Second, it is important that learn grlsvi uses the same noise samples z in across iterations of the for loop of Line 8. To see why, suppose learn grlsvi used independent noise samples in each iteration. Then, when applied to the example of Figure 2, in some iterations, we would be optimistic about the reward at (4, down), while in other iterations, we would be pessimistic about that. Now consider a sample Qθ˜H (1, up) for large H. This sample would be perturbed by a combination of optimistic and pessimistic noise terms influencing assessments at (4, down) to the right. The resulting averaging effect could erode the chances that Qθ˜H(1, up) is optimistic."

He pasted this part carelessly from his thesis. Let me get it. The lines he is talking about do not even exist here.

https://www.youtube.com/watch?v=ck4GixLs4ZQ

Here is his thesis defense.

But let me go for the actual thesis, I do not feel like watching a 2h video.

https://web.stanford.edu/~bvr/pubs/TS_Tutorial.pdf
A Tutorial on Thompson Sampling

Damn, his thesis and this tutorial together are 300 pages.

11am. Putting this and all his papers together, I really have quite a lot of ground to cover. Yeah, I am going to be at this for a while.

Yeah, I understand the uncertainty propagation, but how exactly do I get it to act as an exploration bonus. Will having uncertainty by itself really be enough for truly deep exploration? I can see how it would induce exploration, but would it really be enough to solve the Sea task?

I need just one more extra step. If I could figure it out, I could solve exploration for good.

Let me start with his thesis. I'll do it for a bit and then take a break."