"11:30am. No, I get it. Using local contraction is just not a good idea. It will contract the model uncertainty regardless of whether the pathway has been used. So that only leaves sampling.
It does not have to be from a Gaussian. I could make it absolute Gaussian in order to bias the algorithm towards seeing uncertainty as optimistic. That would be good for uncertainty in the actions. It would still converge to the same thing in the end, but would induce positivity.

Still, I do not like that sort of thing. Noise should be symmetric.

I really can understand why randomized prior functions work so well. They do get around some of the weaknesses of existing method.

11:35am. This thing is really, really difficult to reason about. I am just having a hard time wrapping my head around it. Does it not all reduce to noise in the end?

PSRL, now that I can understand. You sample MDPs and you solve them. I can really see how that would get good exploration. But adding noise to the output? Bootstrapping? My ensemble linking idea?

It feel like I am just stretching it.

I've come to an understanding of what uncertainty is and how it could be propagated through the cost function.

11:45am. Hmmm, I have an idea.

Bootstrapping. Rather than sampling through a Gaussian distribution, why not sample the networks directly?

I am doing a thought experiment and I see that this would actually optimize for epistemic uncertainty rather than A+E. Right now I am imagining a toy dataset of 4 examples (figure 3) and 4 ensembles for which 2 are randomly picked at each pass. I can see that this would force the means towards the correct values. But unlike with Gaussian sampling this would also force the variance towards zero even if there is significant data noise.

Yeah, I understand it now. Without a doubt this it it! I've solved the riddle of how to do exploration.

With just epistemic uncertainty it would be easy to use it as intrinsic reward.

11:55am. There would not be much difference computationally, I'd have to use 10 nets and sample from them rather than 5, but that is not a big deal.

12pm. Actually, rather than fetching the difference directly, you know what would be the easiest thing to do - get rid of the adjoint for the variance and just add it to the outputs. That in fact would have the same behavior as...no, it would bias the value function and I do not want that...

Yeah, it is a bit complex to think of what the best way of taking advantage of the epistemic uncertainty would be. I could add it to the actions that is for sure. And at least half of the benefit of the intrinsic reward could be gotten by adding it to the outputs for the value function. I need to think about that part more in depth.

12:05pm. Let me stop here. I need to eat.

I really haven't been studying much at all in terms of reading the paper, but the ensemble sampling via uniform selection is a great one. It has been worth it to work things out."