"12:15pm. I think that adding the gradient blocked variance to the outputs would work. It is in fact the correct idea. When uncertainty is the same it would even out in the cost function, but when it is not it would lead to it being propagated. And the brilliant thing about split sampling is that it would lead to a biased policy during the episode rather that a noisy one which I'd get with Gaussian sampling.
1:15pm. I am supposed to be resting, but this idea really got me excited, so much that I can't focus on anything else.

Yeah, it is it. I've pretty much solved the exploration problem for model-free RL with just one little thought experiment.

I did not miss it. Not this time. It is my victory.

All I need to do is confirm it on those toy problems and then I can go crush the weak in low stakes poker. After that comes model-based RL with hierarchy.

Well, even crushing the weak will have significant engineering challenges that might take me a year or more to really surmount, but at least now I have something I could use to make a strong player.

It is not like I am trying to beat Dota here. Not at this point anyway.

And it is not the whole of exploration. It is one thing to explore MDPs where you can't assume anything, but true intelligence would know not to jump of a cliff just because it has an exploration bonus.

I am just very lucky that poker is not that kind of game - it is completely artificial and there is not much in terms of priors that I can bake into it. It is also the kind of game where humans would be bad at.

1:40pm. Just like yesterday because of integration, now I am completely distracted. Let me see if I can focus on the thesis just for a bit.

1:50pm. One thing, I am going to have to get rid of the eligibility trace. Monte Carlo is not going to cut it now that I am propagating uncertainties through cost function. I'll just have to go with TD(0). It might as well be like that. It is not like I've any benefit from learnable eligibilities anyway.

2:05pm. I am really not focusing much on the thesis. Well, let me take a break. Who cares about the paper now that I have the solution?

Well, I do and I will going through all the paper properly, but let me take a break here as a reward for my great work. I'll be done with this in a few days anyway."