"3:30pm. Done with chores. There is no way around it, I think I am just going to have to go with the intrinsic motivation idea that I had originally which is to take the difference of scale in the current timestep and the next and use that intrinsic reward.
It should do well enough. The key idea was the uniform sampling without replacement. The realization that it optimizes epistemic uncertainty in its variance is key.

Let me read the thesis. I'll put a few hours into it and then call it a day.

Actually, instead of reading, let me watch 2h video by him.

3:45pm. I just realized, I do not have to strictly get rid of the eligibility trace. I can in fact propagate the intrinsic reward through it just fine. It is no problem.

3:55pm. The quality of audio in his thesis defense is just so bad. My ears are bleeding. Forget it.

4:05pm. I am thinking of how I would handle exploration in the case of Q learning using my method. There is no way around it - just add the scale to values as an exploration bonus. It would be right for action selection even if it is not for calculating targets.

And that is a crucial insight.

If it is right for Q learning then it is right for AC. I am going to add the scale of uncertainty to the actions.

4:10pm. Doing this is only right. It was bothering me a little that I would otherwise not be taking advantage of uncertainty information inherent in the actions.

This would result in a more elaborate method that I had originally wanted, but it is not extremely so. Taking some care with how I handle value predictions is not a huge hurdle to pay.

4:55pm. Done with lunch. It bothers me, it still bothers me.

If uncertainty propagation was right with Gaussian sampling then it should be right with random splitting.

I do not actually think there is a reason why I could not propagate uncertainty using TD learning.

5pm. So it would be a variant of the local contraction idea.

...It would be local contraction done right. I had the sense that relying solely on it would not be a good idea, but maybe this could do it.

In fact, if this is right and it probably is maybe I could get rid of bootstrapping entirely and use local contraction for the actions as well...

5:10pm. No, I'll have to keep bootstrapping in. Uncertainty propagation through the value function is a special case that I need to make everything coherent. Bootstrapping is simply the right way of optimizing for epistemic uncertainty.

The reason why I am wondering whether I could use local contraction everywhere is because in the final episode I will be optimizing with the target scale being zero. But I am being unreasonable to think that is the same as local contraction. The uncertainty is only 0 in the very last episode. Everywhere else the uncertainty will get propagated correctly. So it is hardly the same thing.

5:25pm. I barely read a few dozen pages today. And those what I read I was not really focusing on. But it is just as well, my own thoughts today were definitely more interesting than that guy's thesis.

5:35pm. I am thinking about the whole adding up the predictions thing. Regular ensembles operate by training each ensemble separately.

But probably adding them up and then applying the local contraction would be the equivalent of training them separately. Or I could do the bootstrap thing.

5:45pm. No...

I need to contract the actions. Using just the bootstrap for this sort of task is not right.

But the local contraction is not right either.

Solution - use the Gaussian sampling idea.

Use Gaussian sampling for the actions and uncertainty propagation for the values. If I did it like so I could get rid of the bootstrapping.

5:55pm. Reducing the epistemic uncertainty using solely bootstrapping has the notable weakness of being indirect. I guess it was not a good idea after all.

But uncertainty propagation using TD learning is a really good idea.

6pm. At some point I am simply going to have to take a risk. I cannot do true posterior sampling here.

6:10pm. It will work. As I said training nets individually should be equivalent or almost equivalent to using the mean and output + local contraction of the variance. I simply cannot imagine any other conclusion. I can generalize that local contraction into temporal difference for uncertainty propagation.

6:20pm. I am crazy, obsessed, deranged. I am still thinking about this.

So far I haven't gone for 12h without changing my mind on what I want to do. It is really good that I stopped to think before starting work. If I can go a day or two without changing mind that should be a welcome sign that it is time to start coding.

The plan I have now seems solid enough though. I think it captures most of the salient points of a good exploration method.

6:25pm. Maybe I should call it a day at this point? In the morning when I said I would read manga and rest a bit, I only got through a single damn chapter (of Kaiji) despite the break lasting for an hour. Why?

Because I keep turning the facts over in my head.

6:30pm. Let me just stop. Let me make this my last commit for the day if possible."