"10:30am. Wow, I stayed up in bed for a while today. I guess I'll start out with that break. Before that, let me put the final touch on the idea I had yesterday.
This morning, I thought about temporal difference for uncertainty propagation in the case of MDPs with no terminal states and for a bit I thought that my idea was broken as there would be no way to reduce the uncertainty.

Except that is not true at all - the discount factor is what would lead to uncertainty reduction. In fact, it would lead to a laddered reduction where points away from the starting point have progressively higher uncertainty. This is actually the most ideal possible pattern for exploration. It might be good to keep a separate discount for intrinsic rewards. I might want to make it lower than the extrinsic ones.

10:40am. That is one edge case down. Now the method I discovered is universally applicable. Let me chill for a while. I am not in a hurry here.

I'll go through the thesis and Thompson sampling tutorial today and the rest of the papers tomorrow unless I get another wave of insight, but I do not think that will happen."