"6pm. I am still quite confused as to why BSR works worse than BSP in deep learning - the reason they gave out is so handwavy that I am halfway suspecting implementation issues on their side.
Or maybe the issues is that eq 4 probably does not propagate uncertainty correctly. Eq 5 would do this to an extent just not correctly. Eq 5 is a bit similar to adding uncertainty directly to the value, but who knows what the thing is supposed to do in deep nets. Ian is really grasping the bottom of the barrel with randomized prior functions.

At any rate, that does not matter. I am done with reading. I skimmed the Thomson sampling tutorial as it was not interesting.

6:05pm. There is nothing much to do anymore in terms of reading. I am done.

During the past week I really gave myself time to think about things properly and eventually carved out the correct design for a trully novel RL algorithm. It is my own Christmas present that I gave to myself. It is the power that I gained in return for all my hard work and a chance at a brighter tomorrow.

Tomorrow I will get to work on composing the two updates as planned and after that I will start work on ensembling and the TD+ algorithm."