Right now I am thinking. I've gone over the code for the poker game, but I do not feel the inspiration hitting me exactly.
This is normal. In this kind of situation it won't hit me to begin with. Rather it needs to be induced. I need to start sketching things out in code. That will get the blood flowing through my brain.

It can't be helped. Today at least I will try to break the ice and then go from there. Most of the work will have to be done in the following days.