"6:05pm. Done with lunch. Let me finish the papers and I will read manga for a while or something like that.
Oh, Index is out.

6:15pm. https://scholar.google.com/scholar?as_ylo=2018&hl=hr&as_sdt=0,5&sciodt=0,5&cites=7939374891937847749&scipsc=

This theory of mind stuff should be something I keep up to date with.

6:20pm. https://arxiv.org/abs/1812.01569
Modeling Theory of Mind for Autonomous Agents with Probabilistic Programs

https://arxiv.org/abs/1803.06328
Nesting Probabilistic Programs

7pm. Those papers are beasts. Forget them. I think that for hierarchical learning I will let the field pick up the slack. The same goes for autoencoders. I simply do not have much in the way of good ideas there.

8pm. What was that probabilistic programming hierarchical RL paper? I think it was by Baydin, but now that I am looking at his stuff on Arxiv I can't find it.

http://bayesiandeeplearning.org/2016/papers/BDL_41.pdf
Nested Compiled Inference for Hierarchical Reinforcement Learning

Ah it was this. It is only 4 pages though.

"Here, we show how we can view hierarchical reinforcement learning (RL) [12] as a nested inference scheme in probabilistic programming where we cast policy learning of intrinsic goals as inference on what we call an inner query, which is used inside an outer query that performs inference over sub-goal selection. We then “compile away” the inferential task of the inner query."

Let me read this paper. As I go forward into the future, I am going to have to keep an eye out for how to crack hierarchy. Also I do not understand why regular autoencoders are so noisy just yet. I need figure out what that density estimation thing is.

All this is assuming that I can crack exploration and do not crash and burn again.

8:35pm. https://www.reddit.com/r/MachineLearning/comments/aadhf5/d_are_the_problems_with_training_gans_surmountable/

"Reading [this article](https://medium.com/@jonathan_hui/gan-why-it-is-so-hard-to-train-generative-advisory-networks-819a86b3750b) on why is it so hard to train GANs made me wonder about this. The fact that cost functions might not converge in a minimax game with gradient descent seems especially bad for GANs. That having said, I am not that familiar with them or follow the research on them are so I am asking here.

Will GANs or AEs (variational or otherwise) come out on top in the long term? If so, why?"

8:40pm. https://arxiv.org/abs/1806.02426
Deep Variational Reinforcement Learning for POMDPs

"Lastly, we investigate the influence of the backpropagation length ng on both the RNN and DVRL based policies. While increasing ng universally helps, the key insight here is that a short length ng = 5 (for an average BPTT-length of 2 timesteps) has a stronger negative impact on RNN than on DVRL. This is consistent with our notion that RNN is mainly performing memory based reasoning, for which longer backpropagation-through-time is required: The belief update (2) in DVRL is a one-step update from bt to bt+1, without the need to condition on past actions and observations. The proposal distribution can benefit from extended backpropagation lengths, but this is not necessary. Consequently, this result supports our notion that DVRL relies more on inference computations to update the latent state"

I am going to keep this paper in mind. If my ensembling ideas turn out to work then I am going to have to eventually have to figure out how to deal with belief states and the rest needed for model-based RL. I do not understand that stuff at all at the moment.

https://arxiv.org/abs/1705.10306
Auto-Encoding Sequential Monte Carlo

Ensembling and SMC have a deep connection. I meant to read this paper in my study time, but I do not think I got around to it. I'll just save it for later."