What a hard night. Even though I should be thinking about how to do the game, I just ended stressing myself out about autoencoders. Autoencoders have a deep connection to prediction. In terms of difficulty, reconstructing the state is similar to predicting the next one. I am finally at the stage where I can be really bothered by the fact that regular autoencoders are so blurry. GANs on the other hand have a different kind of possibly unsolvable issues.
And the fact that AEs are so blurry is probably a metalearning problem. It reveals the underlying fact that deep learning is not hierarchical which it will need to be in order to do proper learning.

At this point I have absolutely no idea what the solution to hierarchical learning is. I do not understand how to learn invariants. I just do not know the theory of it. Nor do I know how to make the upper layers direct the lower layers so they learn useful things. Even worse, I understand that it is not just a problem of using better models. I think getting metalearning right is going to involve the optimization pass in some way or another. Better optimizers like KFAC are a kind of metalearning.

This view just makes things more complicated.

A while ago I tried the gradient normalization scheme, but it is not enough. I just need so much more here. And I do not know what it is.

Yesterday I started reading those papers due to restlesness, but I do not think I am going to dive into studies again. I really will try to get some programming done today. I know that I won't find the answer at this point even if I look. For now just getting exploration will be enough. The full theory of learning will have to wait. It is just too difficult at the moment. Exploration being cracked with nested inference is definitely a hint, but I won't be able to do it in a same way. It will have to be a different kind of nested inference.

https://arxiv.org/abs/1806.02426
Deep Variational Reinforcement Learning for POMDPs

I am still not in the programming mood just yet. I will spend some time on this paper even if nothing else.

https://www.reddit.com/r/MachineLearning/comments/aadhf5/d_are_the_problems_with_training_gans_surmountable/

For now let me see what kind of answer I got to the thread I posted on the ML sub yesterday.

...Nothing inspiring. I guess I'll make some replies.