"9:35am. /u/arachnivore: "It depends on what is meant by "are the problems surmountable?" The question is pretty ambiguous. It sounds like OP is saying "it's been shown that there's no guarantee a GAN will converge, does that mean GANs are useless" to which the answer is obviously: No."
/u/abstractcontrol: "They aren't useless if the goal is to create realistic images, but the stability is an issue if you'd want to use GANs for state prediction (world modeling) in a RL agent. I see prediction and reconstruction quite similar to each other as you need to be good at a lot of the same kinds of things to do both tasks.

The current wave of learning while deep is not necessarily hierarchical. It is still a mystery of how learn invariants and how to make layers cooperate in a hierarchical fashion to do that. GANs are hot right now, but in a way they are the exact opposite of hierarchy as they work on adversity rather than cooperation.

I made the question ambiguous on purpose as I did not want to bias the responses by giving my view first. I even used 'surmountable' as opposed to 'insurmountable'.

I'd speculate that how well autoencoders work is actually a good barometer of how close we are to cracking learning as a whole.

> to which the answer is obviously: No.

It is far from obvious that the answer is no. Autoencoders would be better served with an aggressive attack than an aggressive defense."

9:40am. This is good. Let me chill a bit and then I'll read the RL paper. I'll try to get some programming done after the break."